+++ Changelog.md
@@ -1,8 +1,11 @@
 Changelog
 =========
 
+# Version 1.2.7
+- Attempt to warn when regions such as `_HOME` or `_CODE` have overflowed ROM0 and rom size is larger than 32k
+
 # Version 1.2.6
-- Added support for GBDK's flavor of Game Gear/SMS (`-p:SMS_GG"`) .noi/.map files (with LIT and CODE areas)
+- Added support for GBDK's flavor of Game Gear/SMS (`-p:SMS_GG"`) .noi/.map files (with `_LIT` and `_CODE` areas)
 - Fix: free % + used% did not always equal 100%
 
 # Version 1.2.5